Logger And ErrorList: Complete

+++ Epam.FitnessCenter/Epam.FitnessCenter.BLL/HallLogic.cs
@@ -1,9 +1,11 @@
-﻿using Epam.FitnessCenter.DAL.Interface;
+﻿using Epam.FitnessCenter.CustomException;
+using Epam.FitnessCenter.DAL.Interface;
 using Epam.FitnessCenter.Entities;
 using System.Collections.Generic;
 
 namespace Epam.FitnessCenter.BLL.Interface
 {
+    //TODO Logger
     public class HallLogic : IHallLogic
     {
         private IHallDao _hallDao;
@@ -13,10 +15,27 @@ namespace Epam.FitnessCenter.BLL.Interface
             _hallDao = hallDao;
         }
 
-        public void Add(Hall hall, out IEnumerable<Error> listError)
+        public void Add(Hall hall, out ICollection<Error> listError)
         {
             listError = new List<Error>();
-            _hallDao.Add(hall);
+            try
+            {
+                _hallDao.Add(hall);
+            }
+            catch (UniqueIdentifierException ex)
+            {
+                listError.Add(new Error
+                {
+                    Message = ex.Message
+                });
+            }
+            catch
+            {
+                listError.Add(new Error
+                {
+                    Message = "Failed to add hall, please try again"
+                });
+            }
         }
 
         public IEnumerable<Hall> GetAll()
@@ -29,10 +48,28 @@ namespace Epam.FitnessCenter.BLL.Interface
             return _hallDao.GetById(id);
         }
 
-        public void Remove(int id, out IEnumerable<Error> listError)
+        public void Remove(int id, out ICollection<Error> listError)
         {
             listError = new List<Error>();
-            _hallDao.Remove(id);
+            try
+            {
+                if (_hallDao.GetById(id) == null)
+                {
+                    listError.Add(new Error
+                    {
+                        Message = "Hall won't find"
+                    });
+                    return;
+                }
+                _hallDao.Remove(id);
+            }
+            catch
+            {
+                listError.Add(new Error
+                {
+                    Message = "Internal error, try again"
+                });
+            }
         }
     }
 }
