+++ Epam.FitnessCenter/Epam.FitnessCenter.BLL/UserLogic.cs
@@ -4,8 +4,6 @@ using Epam.FitnessCenter.DAL.Interface;
 using Epam.FitnessCenter.Entities;
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.Runtime.CompilerServices;
 
 namespace Epam.FitnessCenter.BLL
 {
@@ -33,12 +31,19 @@ namespace Epam.FitnessCenter.BLL
                     Message = ex.Message
                 });
             }
+            catch (RoleUndefinedException)
+            {
+                listError.Add(new Error
+                {
+                    Message = "Registration failed, please try again"
+                });
+            }
             catch
             {
                 listError.Add(new Error
                 {
-                    Message = "User not added, internal error occurred, please try again"
-                }); ;
+                    Message = "Something went wrong"
+                });
             }
 
         }
@@ -56,13 +61,53 @@ namespace Epam.FitnessCenter.BLL
         public void Remove(int id, out ICollection<Error> listError)
         {
             listError = new List<Error>();
-            _userDao.Remove(id);
+            try
+            {
+                if (_userDao.GetById(id) == null)
+                {
+                    listError.Add(new Error
+                    {
+                        Message = "User won't find"
+                    });
+                    return;
+                }
+
+                _userDao.Remove(id);
+            }
+            //TODO Logger
+            catch
+            {
+                listError.Add(new Error
+                {
+                    Message = "Internal error, try again"
+                });
+            }
         }
 
         public void Update(User user, out ICollection<Error> listError)
         {
             listError = new List<Error>();
-            _userDao.Update(user);
+            try
+            {
+                if (_userDao.GetById(user.Id) == null)
+                {
+                    listError.Add(new Error
+                    {
+                        Message = "User won't find"
+                    });
+                    return;
+                }
+
+                _userDao.Update(user);
+            }
+            //TODO Logger
+            catch (RoleUndefinedException ex)
+            {
+                listError.Add(new Error
+                {
+                    Message = ex.Message
+                });
+            }
         }
     }
 }
